Move search under its own package
Move highlevel classes under highlevel package as well
