+++ client/highlevel-rest/src/test/java/org/elasticsearch/client/highlevel/XContentAccessorTests.java
@@ -16,8 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.elasticsearch.client;
+package org.elasticsearch.client.highlevel;
 
+import org.elasticsearch.client.highlevel.XContentAccessor;
 import org.elasticsearch.common.xcontent.XContentBuilder;
 import org.elasticsearch.common.xcontent.XContentFactory;
 import org.elasticsearch.common.xcontent.XContentType;
